fix(auth): mypy/ruff 검증 중 발견된 타입·린트 이슈 수정 (TASK-32:6)
- KC_HTTP_TIMEOUT_SECONDS를 float에서 int로 변경: python-keycloak의
  KeycloakOpenID/KeycloakAdmin timeout 파라미터가 int만 받아 mypy가
  실패했다.
- 신규 회귀 테스트의 access_token 리터럴 비교에 ruff S105(하드코딩 비밀번호
  오탐) noqa를 추가해 베이스라인(31건) 대비 신규 위반이 없도록 했다.

uv run ruff check .와 uv run mypy .를 이 변경 전후로 비교해, 기존에도
존재하던 베이스라인 위반(ruff 31건, mypy 10건, pytest 5건 admin_crud
실패)과 무관하게 이 브랜치가 새 위반을 추가하지 않음을 확인했다.

+++ app/config/config.py
@@ -84,7 +84,8 @@ class Config:
     KC_ADMIN_ROLE = os.getenv("KC_ADMIN_ROLE", "global_admin")
     # 카카오 응답 예산(5초)을 넘기지 않도록 Keycloak 클라이언트 HTTP 타임아웃을 제한한다.
     # DB 조회·JWT 디코딩 등 나머지 처리 시간을 남겨두기 위해 예산보다 짧게 잡는다.
-    KC_HTTP_TIMEOUT_SECONDS = float(os.getenv("KC_HTTP_TIMEOUT_SECONDS", "3"))
+    # python-keycloak의 timeout 파라미터가 int만 받으므로 초 단위 정수로 둔다.
+    KC_HTTP_TIMEOUT_SECONDS = int(os.getenv("KC_HTTP_TIMEOUT_SECONDS", "3"))
 
     ADMIN_PANEL_ENABLED = os.getenv("ADMIN_PANEL_ENABLED", "true").lower() == "true"
     ADMIN_SESSION_TTL_SECONDS = int(os.getenv("ADMIN_SESSION_TTL_SECONDS", "3600"))
